ManagedStatic: remove from GDBRegistrationListener
An earlier version of this change originally landed as part of
e6248b566b0b8836a766374fefbac754295313e3 (D129120), which caused a
Fuchsia buildbot regression in ExecutionEngine tests.

Careful review suggests that the issue was that in the earlier version,
the destructor of the JITDebugLock was run before the destructor of
GDBJITRegistrationListener. The new version of the change moves the lock
to a member variable of the (singleton!) GDBJITRegistartionListener so
that destructors are run in the right order.