[clang-tidy] Rename the make-confusable-table executable
Rename it to clang-tidy-confusable-chars-gen, to make its role
clearer in a wider context.

In cross builds, the caller might want to provide this tool
externally (to avoid needing to rebuild it in the cross build).
In such a case, having the tool properly namespaced makes its role
clearer.

This matches how the clang-pseudo-gen tool was renamed in
9fe139d866355a492fd6f7af0258737aac7650d1 / D126725.

Differential Revision: https://reviews.llvm.org/D129798

(cherry picked from commit 18b4a8bcf3553174f770f09528c9bd01c8cebfe7)