Re-re-apply 0baaa75683d5, Add a shared-memory based orc::MemoryMapper...
...with more fixes.

The original patch was reverted in 8f1c88dea51 due to bot failures caused by
a missing dependence on librt. That issue was fixed in 3eb8bdc75b, but that
commit also broke sanitizer bots due to a bug in SimplePackedSerialization:
empty ArrayRef<char>s triggered a zero-byte memcpy from a null source. The
ArrayRef<char> serialization issue was fixed in 358938d5ed, and this patch has
also been updated with a new custom SharedMemorySegFinalizeRequest message that
should avoid serializing empty ArrayRefs in the first place.

https://reviews.llvm.org/D128544